pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "slop-doc"
version = "0.1.0"
description = "Static documentation generator for Python projects"
readme = "README.md"
authors = [{name = "Artem"}]
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "pyyaml",
    "markdown",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov"]

[project.scripts]
slop-doc = "slop_doc.builder:main"

[tool.setuptools.package-data]
slop_doc = ["defaults/*.css", "defaults/*.js", "defaults/templates/*.dtmpl"]

[tool.setuptools.packages.find]
where = ["."]
